✨ Feat: ProductDetail 컴포넌트 분리와 해당 컴포넌트를 담을 Page.tsx 생성
